+++ connection.impl/src/main/java/com/ericsson/bss/cassandra/ecchronos/connection/impl/LocalNativeConnectionProvider.java
@@ -23,6 +23,7 @@ import com.datastax.oss.driver.api.core.config.ProgrammaticDriverConfigLoaderBui
 import com.datastax.oss.driver.api.core.metadata.EndPoint;
 import com.datastax.oss.driver.api.core.metadata.Node;
 import com.datastax.oss.driver.api.core.metadata.NodeStateListener;
+import com.datastax.oss.driver.api.core.metadata.schema.KeyspaceMetadata;
 import com.datastax.oss.driver.api.core.metadata.schema.SchemaChangeListener;
 import com.datastax.oss.driver.api.core.metrics.DefaultNodeMetric;
 import com.datastax.oss.driver.api.core.metrics.DefaultSessionMetric;
@@ -39,6 +40,8 @@ import org.slf4j.LoggerFactory;
 import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
+import java.util.Optional;
+import java.util.Map;
 
 public final class LocalNativeConnectionProvider implements NativeConnectionProvider
 {
@@ -97,7 +100,6 @@ public final class LocalNativeConnectionProvider implements NativeConnectionProv
     {
         return myRemoteRouting;
     }
-
     @Override
     public void close()
     {
@@ -122,6 +124,7 @@ public final class LocalNativeConnectionProvider implements NativeConnectionProv
         private SchemaChangeListener mySchemaChangeListener = null;
         private NodeStateListener myNodeStateListener = null;
         private MeterRegistry myMeterRegistry = null;
+        private String myRepairHistoryKeyspace = "ecchronos";
 
         public final Builder withLocalhost(final String localhost)
         {
@@ -176,7 +179,11 @@ public final class LocalNativeConnectionProvider implements NativeConnectionProv
             myMeterRegistry = meterRegistry;
             return this;
         }
-
+        public final Builder withRepairHistoryKeyspace(final String repairHistoryKeyspace)
+        {
+            myRepairHistoryKeyspace = repairHistoryKeyspace;
+            return this;
+        }
         public final LocalNativeConnectionProvider build()
         {
             CqlSession session = createSession(this);
@@ -242,6 +249,15 @@ public final class LocalNativeConnectionProvider implements NativeConnectionProv
                 {
                     if (node.getEndPoint().equals(contactEndPoint))
                     {
+                        // check to ensure ecchronos keyspace is replicated correctly
+                        Optional<KeyspaceMetadata> keyspaceMetadata =
+                                session.getMetadata().getKeyspace(builder.myRepairHistoryKeyspace);
+                        Map<String, String> replication = keyspaceMetadata.get().getReplication();
+                        if (!replication.containsKey(node.getDatacenter()))
+                        {
+                            throw new IllegalStateException("Keyspace " + builder.myRepairHistoryKeyspace
+                                    + " not replicated on local node.");
+                        }
                         return new InitialContact(node.getDatacenter(), node.getHostId());
                     }
                 }